refactor: change float to Integer in price

+++ app/models/product.py
@@ -1,5 +1,5 @@
 from typing import List, TYPE_CHECKING
-from sqlalchemy import String, Text, Float
+from sqlalchemy import String, Text, Integer
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.db.session import Base
@@ -13,6 +13,6 @@ class Product(Base):
     id: Mapped[int] = mapped_column(primary_key=True, index=True)
     name: Mapped[str] = mapped_column(String(255), unique=True, index=True)
     description: Mapped[str] = mapped_column(Text, nullable=True)
-    price: Mapped[float] = mapped_column(Float, default=0.0)
+    price: Mapped[int] = mapped_column(Integer, default=0)
 
     recipe_items: Mapped[List["RecipeItem"]] = relationship(back_populates="product", cascade="all, delete-orphan")
